+++ git_improved/templates/__init__.py
@@ -1,0 +1,5 @@
+from .python_package import PythonPackageTemplate
+
+TEMPLATES = {
+    'python-package': PythonPackageTemplate
+}